+++ src/platform_impl/linux/x11/window.rs
@@ -454,6 +454,21 @@ impl UnownedWindow {
             );
             leap!(result).ignore_error();
 
+            // Set the transient parent if we don't have a parent already.
+            if event_loop.root == root {
+                let owner = pl_attribs.owner.map(|owner| match owner {
+                    RawWindowHandle::Xlib(x) => x.window,
+                    RawWindowHandle::Xcb(x) => x.window as u64,
+                    raw => unreachable!("Invalid raw window handle {raw:?} on X11"),
+                });
+
+                if let Some(owner) = owner {
+                    unsafe {
+                        (xconn.xlib.XSetTransientForHint)(xconn.display, window.xwindow, owner);
+                    }
+                }
+            }
+
             // Set visibility (map window)
             if window_attrs.visible {
                 leap!(xconn.xcb_connection().map_window(window.xwindow)).ignore_error();